chore: correct type of the function's input parameters

+++ src/grammar/grammar_analysis_utils.c
@@ -1,6 +1,6 @@
 #include "grammar_analysis_utils.h"
 
-MRGrammar_t get_alpha_grammar(MRGrammarType curGrammar, int n_par, int n_bra,
+MRGrammar_t get_alpha_grammar(MRGrammarType curGrammar, int64_t n_par, int64_t n_bra,
 							  bool has_normal) {
 	switch (curGrammar) {
 		case PARITY:
@@ -20,7 +20,7 @@ MRGrammar_t get_alpha_grammar(MRGrammarType curGrammar, int n_par, int n_bra,
 	}
 }
 
-MRGrammar_t get_beta_grammar(MRGrammarType curGrammar, int n_par, int n_bra,
+MRGrammar_t get_beta_grammar(MRGrammarType curGrammar, int64_t n_par, int64_t n_bra,
 							 bool has_normal) {
 	switch (curGrammar) {
 		case PARITY:
