+++ src/grammar/grammar_analysis_utils.h
@@ -2,8 +2,8 @@
 
 #include "grammar.h"
 
-MRGrammar_t get_alpha_grammar(MRGrammarType curGrammar, int n_par, int n_bra,
+MRGrammar_t get_alpha_grammar(MRGrammarType curGrammar, int64_t n_par, int64_t n_bra,
 							  bool has_normal);
 
-MRGrammar_t get_beta_grammar(MRGrammarType curGrammar, int n_par, int n_bra,
+MRGrammar_t get_beta_grammar(MRGrammarType curGrammar, int64_t n_par, int64_t n_bra,
 							 bool has_normal);
